+++ GaudiKernel/include/GaudiKernel/SmartRefBase.h
@@ -1,5 +1,5 @@
 /***********************************************************************************\
-* (c) Copyright 1998-2019 CERN for the benefit of the LHCb and ATLAS collaborations *
+* (c) Copyright 1998-2021 CERN for the benefit of the LHCb and ATLAS collaborations *
 *                                                                                   *
 * This software is distributed under the terms of the Apache version 2 licence,     *
 * copied verbatim in the file "LICENSE".                                            *
@@ -111,7 +111,7 @@ public:
   const std::string& path() const;
 
 public:
-  int objectType() const { return m_type; }
+  long objectType() const { return m_type; }
   /// Setup smart reference when reading. Must be allowed from external sources.
   void set( DataObject* pObj, long hint_id, long link_id );
 };
